Combine project and static metadata
