+++ src/client/actions.ts
@@ -1,20 +1,13 @@
 import { ActionType } from "typesafe-actions";
 
-import * as projectActions from "./actions/project";
+import * as projectDataActions from "./actions/projectData";
 import * as projectsActions from "./actions/projects";
 import * as regionConfigActions from "./actions/regionConfig";
-import * as staticMetadataActions from "./actions/staticMetadata";
 import * as userActions from "./actions/user";
 
-export type ProjectAction = ActionType<typeof projectActions>;
+export type ProjectDataAction = ActionType<typeof projectDataActions>;
 export type ProjectsAction = ActionType<typeof projectsActions>;
 export type RegionConfigAction = ActionType<typeof regionConfigActions>;
-export type StaticMetadataAction = ActionType<typeof staticMetadataActions>;
 export type UserAction = ActionType<typeof userActions>;
 
-export type Action =
-  | UserAction
-  | RegionConfigAction
-  | StaticMetadataAction
-  | ProjectAction
-  | ProjectsAction;
+export type Action = UserAction | RegionConfigAction | ProjectDataAction | ProjectsAction;
