--- src/client/actions/staticMetadata.ts
@@ -1,8 +1,0 @@
-import { createAction } from "typesafe-actions";
-import { IStaticMetadata, S3URI } from "../../shared/entities";
-
-export const staticMetadataFetch = createAction("StaticMetadata fetch")<S3URI>();
-export const staticMetadataFetchSuccess = createAction("StaticMetadata fetch success")<
-  IStaticMetadata
->();
-export const staticMetadataFetchFailure = createAction("StaticMetadata fetch failure")<string>();
